+++ angular.json
@@ -2,7 +2,9 @@
   "$schema": "./node_modules/@angular/cli/lib/config/schema.json",
   "cli": {
     "analytics": "a3fcfd43-a294-4b2d-ae71-a7a806f66438",
-    "schematicCollections": ["@ngrx/schematics"]
+    "schematicCollections": [
+      "@ngrx/schematics"
+    ]
   },
   "version": 1,
   "newProjectRoot": "projects",
@@ -53,17 +55,30 @@
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
-            "localize": ["en-CH"],
+            "localize": [
+              "en-CH"
+            ],
             "outputPath": "dist/cloud-portal",
             "index": "src/index.html",
             "main": "src/main.ts",
-            "polyfills": "src/polyfills.ts",
+            "polyfills": [
+              "src/structuredClone.polyfill.ts",
+              "zone.js"
+            ],
             "tsConfig": "tsconfig.app.json",
             "inlineStyleLanguage": "scss",
-            "assets": ["src/manifest.webmanifest", "src/config.json", "src/assets"],
-            "styles": ["src/styles.scss"],
+            "assets": [
+              "src/manifest.webmanifest",
+              "src/config.json",
+              "src/assets"
+            ],
+            "styles": [
+              "src/styles.scss"
+            ],
             "scripts": [],
-            "allowedCommonJsDependencies": ["fast-sha256"]
+            "allowedCommonJsDependencies": [
+              "fast-sha256"
+            ]
           },
           "configurations": {
             "production": {
@@ -143,7 +158,10 @@
         "lint": {
           "builder": "@angular-eslint/builder:lint",
           "options": {
-            "lintFilePatterns": ["src/**/*.ts", "src/**/*.html"]
+            "lintFilePatterns": [
+              "src/**/*.ts",
+              "src/**/*.html"
+            ]
           }
         },
         "cypress-run": {
