+++ tsconfig.app.json
@@ -3,8 +3,15 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "outDir": "./out-tsc/app",
-    "types": []
+    "types": [
+      "@angular/localize"
+    ]
   },
-  "files": ["src/main.ts", "src/polyfills.ts"],
-  "include": ["src/**/*.d.ts"]
+  "files": [
+    "src/main.ts",
+    "src/structuredClone.polyfill.ts"
+  ],
+  "include": [
+    "src/**/*.d.ts"
+  ]
 }
