+++ tsconfig.spec.json
@@ -4,8 +4,13 @@
   "compilerOptions": {
     "esModuleInterop": true,
     "outDir": "./out-tsc/spec",
-    "types": ["jest"]
+    "types": [
+      "jest"
+    ]
   },
-  "files": ["src/polyfills.ts"],
-  "include": ["src/**/*.spec.ts", "src/**/*.d.ts"]
+  "files": [],
+  "include": [
+    "src/**/*.spec.ts",
+    "src/**/*.d.ts"
+  ]
 }
